Save model performance tables and include PMIDs
- To allow inspection of publications of different scores.
- Also allows later generation of figures for more customization.
- Fixes an indexing error.

+++ abstract2gene/model/_models.py
@@ -125,7 +125,7 @@ class Model(nnx.Module):
         in_dataset = np.isin(temp_names, ds_names)
         missing_count = 0
 
-        indices = np.zeros(temp_names.shape[0])
+        indices = np.zeros((temp_names.shape[0],), dtype=int)
         for i in range(temp_names.shape[0]):
             if in_dataset[i]:
                 indices[i] = name2index(temp_names[i])
